finished test 'gets paid $12/hour from start-time to bedtime'

+++ BabySitterKataTests/BabySitterKataTests.swift
@@ -71,7 +71,7 @@ class BabySitterKataTests: XCTestCase {
       let endTime: NSDate = dateFormatter.dateFromString("20:00:00"),
       let bedTime: NSDate = dateFormatter.dateFromString("20:00:00") {
         
-        let correctPay: Int = 12 * 3 // $12/hr for 3hr
+        let correctPay: Double = 12 * 3 // $12/hr for 3hr
         
         XCTAssert(babySitter.calculatePayFromStartTimeToBedTime(startTime, endTime: endTime, bedTime: bedTime) == correctPay, "BabySitter calculatePayFromStartTimeToBedTime incorrect.")
     }
@@ -80,7 +80,7 @@ class BabySitterKataTests: XCTestCase {
       let endTime: NSDate = dateFormatter.dateFromString("19:00:00"),
       let bedTime: NSDate = dateFormatter.dateFromString("20:00:00") {
         
-        let correctPay: Int = 12 * 2 // $12/hr for 2hr
+        let correctPay: Double = 12 * 2 // $12/hr for 2hr
         
         XCTAssert(babySitter.calculatePayFromStartTimeToBedTime(startTime, endTime: endTime, bedTime: bedTime) == correctPay, "BabySitter calculatePayFromStartTimeToBedTime incorrect.")
     }
@@ -89,7 +89,7 @@ class BabySitterKataTests: XCTestCase {
       let endTime: NSDate = dateFormatter.dateFromString("22:00:00"),
       let bedTime: NSDate = dateFormatter.dateFromString("20:00:00") {
         
-        let correctPay: Int = 12 * 3 // $12/hr for 3hr
+        let correctPay: Double = 12 * 3 // $12/hr for 3hr
         
         XCTAssert(babySitter.calculatePayFromStartTimeToBedTime(startTime, endTime: endTime, bedTime: bedTime) == correctPay, "BabySitter calculatePayFromStartTimeToBedTime incorrect.")
     }
